# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OnboardYou Config API",
    "description": "REST API for managing OnboardYou ETL pipeline configurations.\n\nSupports CRUD operations on pipeline configs stored in DynamoDB and \ndry-run validation of pipeline manifests via column propagation.",
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/config": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "summary": "GET /config",
        "description": "List all pipeline configurations owned by the caller's organization.\nThe organization is resolved from the caller's JWT claims.",
        "operationId": "list_configs",
        "responses": {
          "200": {
            "description": "List of pipeline configurations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PipelineConfig"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/config/{customer_company_id}": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "summary": "GET /config/{customerCompanyId}",
        "description": "Retrieve the pipeline configuration for a given customer company.\nThe organization is resolved from the caller's JWT claims.",
        "operationId": "get_config",
        "parameters": [
          {
            "name": "customer_company_id",
            "in": "path",
            "description": "Unique identifier for the customer company",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pipeline configuration found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Configuration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Configuration"
        ],
        "summary": "PUT /config/{customerCompanyId}",
        "description": "Update an existing pipeline configuration. Validates the pipeline manifest,\npersists to DynamoDB, and updates the EventBridge schedule.\nThe organization is resolved from the caller's JWT claims.",
        "operationId": "update_config",
        "parameters": [
          {
            "name": "customer_company_id",
            "in": "path",
            "description": "Unique identifier for the customer company",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Pipeline configuration to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Configuration updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Configuration"
        ],
        "summary": "POST /config/{customerCompanyId}",
        "description": "Create a new pipeline configuration. Validates the pipeline manifest,\npersists to DynamoDB, and creates an EventBridge schedule.\nThe organization is resolved from the caller's JWT claims.",
        "operationId": "create_config",
        "parameters": [
          {
            "name": "customer_company_id",
            "in": "path",
            "description": "Unique identifier for the customer company",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Pipeline configuration to create",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Configuration created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Configuration"
        ],
        "summary": "DELETE /config/{customerCompanyId}",
        "description": "Delete a pipeline configuration and its associated EventBridge schedule.\nThe organization is resolved from the caller's JWT claims.",
        "operationId": "delete_config",
        "parameters": [
          {
            "name": "customer_company_id",
            "in": "path",
            "description": "Unique identifier for the customer company",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Configuration deleted successfully"
          },
          "401": {
            "description": "Unauthorized — missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/config/{customer_company_id}/generate-plan": {
      "post": {
        "tags": [
          "Configuration"
        ],
        "summary": "POST /config/{customerCompanyId}/generate-plan",
        "description": "Trigger async AI plan generation for a pipeline configuration.\nSets the generation status to InProgress, sends an SQS message to the\netl-trigger lambda, and returns 202 immediately. The frontend polls\n`GET /config/{id}` until `planSummary.generationStatus` is `Completed`.",
        "operationId": "generate_plan",
        "parameters": [
          {
            "name": "customer_company_id",
            "in": "path",
            "description": "Unique identifier for the customer company",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Source system information for plan generation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Plan generation started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratePlanResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Configuration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/config/{customer_company_id}/validate": {
      "post": {
        "tags": [
          "Validation"
        ],
        "summary": "POST /config/{customerCompanyId}/validate",
        "description": "Dry-run column propagation: parses the pipeline manifest, builds every\naction, and folds `calculate_columns` through the chain. Returns the\ncolumn set after each step without executing any real transformations.",
        "operationId": "validate_config",
        "parameters": [
          {
            "name": "customer_company_id",
            "in": "path",
            "description": "Unique identifier for the customer company",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Pipeline configuration to validate (only the pipeline field is used)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Validation result with per-step column snapshots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            }
          },
          "400": {
            "description": "Validation error — misconfigured action or schema conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionConfig": {
        "type": "object",
        "description": "Configuration for a single action in the pipeline",
        "required": [
          "id",
          "action_type",
          "config"
        ],
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/ActionType",
            "description": "Factory key — selects the Rust implementation (e.g. `ActionType::CsvHrisConnector`)"
          },
          "config": {
            "$ref": "#/components/schemas/ActionConfigPayload",
            "description": "Action-specific configuration (shape depends on action_type)\n\nDeserialized into the concrete typed variant matching `action_type`\nvia the custom `Deserialize` impl below."
          },
          "disabled": {
            "type": "boolean",
            "description": "When `true`, the ETL engine skips this action at runtime.\nUsed by the plan summary UI to toggle features on/off."
          },
          "id": {
            "type": "string",
            "description": "Unique identifier for this pipeline step"
          }
        }
      },
...
      "ConfigRequest": {
        "type": "object",
        "description": "Request body for `POST /config/{id}` and `PUT /config/{id}`.\n\nClients send only the fields they own; the server stamps identity\nand timestamps before persisting.",
        "required": [
          "name",
          "cron",
          "pipeline"
        ],
        "properties": {
          "cron": {
            "type": "string",
            "description": "EventBridge-compatible schedule expression (cron or rate)"
          },
          "image": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional image/icon for the pipeline"
          },
          "name": {
            "type": "string",
            "description": "Name of the pipeline"
          },
          "pipeline": {
            "$ref": "#/components/schemas/Manifest",
            "description": "The full ETL pipeline manifest"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "JSON error envelope returned to clients.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error description"
          }
        }
      },
...
      "GeneratePlanRequest": {
        "type": "object",
        "description": "Request body for `POST /config/{id}/generate-plan`.",
        "required": [
          "sourceSystem"
        ],
        "properties": {
          "sourceSystem": {
            "type": "string",
            "description": "Source system name — \"Workday\" or \"CSV\""
          }
        }
      },
      "GeneratePlanResponse": {
        "type": "object",
        "description": "Response body for `POST /config/{id}/generate-plan` (202 Accepted).",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "description": "Current generation status"
          }
        }
      },
...
      "Manifest": {
        "type": "object",
        "description": "Version of the manifest schema",
        "required": [
          "version",
          "actions"
        ],
        "properties": {
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActionConfig"
            },
            "description": "Ordered list of pipeline actions"
          },
          "version": {
            "type": "string",
            "description": "Schema version (e.g. \"1.0\")"
          }
        }
      },
...
      "PipelineConfig": {
        "type": "object",
        "description": "The pipeline config as stored in DynamoDB and exchanged via the API.\n\n```json\n{\n  \"name\": \"Acme Onboarding Pipeline\",\n  \"cron\": \"rate(1 hour)\",\n  \"organizationId\": \"acme-corp\",\n  \"customerCompanyId\": \"customer-123\",\n  \"lastEdited\": \"2026-02-09T12:00:00Z\",\n  \"pipeline\": { \"version\": \"1.0\", \"actions\": [...] }\n}\n```",
        "required": [
          "name",
          "cron",
          "organizationId",
          "customerCompanyId",
          "pipeline"
        ],
        "properties": {
          "cron": {
            "type": "string",
            "description": "EventBridge-compatible schedule expression (cron or rate)"
          },
          "customerCompanyId": {
            "type": "string",
            "description": "Unique identifier for the customer company (sort key)"
          },
          "image": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional image/icon for the pipeline"
          },
          "lastEdited": {
            "type": "string",
            "description": "ISO 8601 timestamp of last edit — set by the server"
          },
          "name": {
            "type": "string",
            "description": "Name of the pipeline"
          },
          "organizationId": {
            "type": "string",
            "description": "Unique identifier for the organization (partition key)"
          },
          "pipeline": {
            "$ref": "#/components/schemas/Manifest",
            "description": "The full ETL pipeline manifest (passed through to the ETL Lambda)"
          },
          "planSummary": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PlanSummary",
                "description": "AI-generated plan summary — cached from the plan generation engine.\n`None` until plan generation has been triggered."
              }
            ]
          }
        }
      },
      "PlanFeature": {
        "type": "object",
        "description": "A single feature card in the plan summary UI.\n\nEach feature references one or more manifest actions via `action_ids`.\nToggling a feature flips the `disabled` flag on those actions.",
        "required": [
          "id",
          "icon",
          "label",
          "description",
          "actionIds"
        ],
        "properties": {
          "actionIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "References to the manifest action IDs this feature maps to"
          },
          "description": {
            "type": "string",
            "description": "Description of what this feature does"
          },
          "icon": {
            "type": "string",
            "description": "Icon name for the UI (e.g. \"calendar\", \"users\", \"shield\")"
          },
          "id": {
            "type": "string",
            "description": "Unique identifier for this feature (e.g. \"sync_start_dates\")"
          },
          "label": {
            "type": "string",
            "description": "Human-readable label (e.g. \"Sync Start Dates\")"
          }
        }
      },
...
      "PlanSummary": {
        "type": "object",
        "description": "AI-generated plan summary cached on `PipelineConfig`.\n\nProvides a human-readable description of the pipeline plus\ntoggleable features that map back to manifest actions.",
        "required": [
          "headline",
          "description",
          "features",
          "preview",
          "generationStatus"
        ],
        "properties": {
          "description": {
            "type": "string",
            "description": "Multi-sentence description of the pipeline"
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanFeature"
            },
            "description": "Toggleable features — each references manifest action IDs"
          },
          "generationStatus": {
            "$ref": "#/components/schemas/SchemaGenerationStatus",
            "description": "Current generation status (InProgress, Completed, Failed)"
          },
          "headline": {
            "type": "string",
            "description": "One-line headline (e.g. \"Here's the plan to connect Workday to your App.\")"
          },
          "preview": {
            "$ref": "#/components/schemas/PlanPreview",
            "description": "Synthetic before/after preview data"
          }
        }
      },
...
      "SchemaDiff": {
        "type": "object",
        "description": "Result of diffing `final_columns` against the egress schema.\n\nShows which pipeline columns map to destination fields, and which are\nunmatched on either side.",
        "required": [
          "mapped",
          "unmapped_source",
          "unmapped_target"
        ],
        "properties": {
          "mapped": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnMapping"
            },
            "description": "Columns present in the pipeline that have a mapping in the egress schema"
          },
          "unmapped_source": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Pipeline columns with no egress mapping"
          },
          "unmapped_target": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Egress schema fields with no matching pipeline column"
          }
        }
      },
      "SchemaGenerationStatus": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "notStarted"
            ]
          },
          {
            "type": "string",
            "enum": [
              "inProgress"
            ]
          },
          {
            "type": "string",
            "enum": [
              "completed"
            ]
          },
          {
            "type": "object",
            "required": [
              "failed"
            ],
            "properties": {
              "failed": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "StepValidation": {
        "type": "object",
        "description": "Result of validating a single step in the pipeline.",
        "required": [
          "action_id",
          "action_type",
          "columns_after"
        ],
        "properties": {
          "action_id": {
            "type": "string",
            "description": "Action id from the manifest."
          },
          "action_type": {
            "type": "string",
            "description": "Action type (e.g. `csv_hris_connector`, `drop_column`)."
          },
          "columns_after": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Columns present *after* this step completes."
          }
        }
      },
      "ValidationResult": {
        "type": "object",
        "description": "Overall validation result for the entire pipeline.",
        "required": [
          "steps",
          "final_columns"
        ],
        "properties": {
          "final_columns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Final column set after the last step."
          },
          "schema_diff": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SchemaDiff",
                "description": "Schema diff against the egress model (if an egress action is present)."
              }
            ]
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StepValidation"
            },
            "description": "Per-step column snapshots (in execution order)."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Cognito JWT — the `custom:organizationId` claim is extracted by the Lambda authorizer and injected into the request context."
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}